% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Release_kinetics.R
\name{rk_second}
\alias{rk_second}
\title{Fitting of release kinetics data to the second-order equation}
\usage{
rk_second(Kt = Kt, Time = Time,...)
}
\arguments{
\item{Kt}{Cumulative amount of element released (mass or mole of element per unit mass of soil, e.g., mg/kg or mmol/kg) after time ‘t’}

\item{Time}{The time intervals used in the kinetics study (user can choose the unit as per convenience)}

\item{...}{Any other argument that can be passed to base plot}
}
\value{
\itemize{
\item R2: Coefficient of determination (more the value, better the fit)
\item SEE: Standard error of estimate (less the value, better the fit)
\item a: Equation constant
\item b: Equation constant (overall release rate coefficient)
}
}
\description{
The function fits the cumulative amount of an element released from soil with a given extractant over a certain time period to the second-order equation. It generates the coefficient of determination (R-squared) and standard error of estimate to show the goodness of fit. The parameters of second-order equation, i.e., ‘Kmax’ and ‘t1/2’ can be obtained from this function.
}
\details{
The second-order equation is expressed as t ÷ Kt = (t ÷ Kmax) + (t1/2 ÷ Kmax) (Grimme 1980; Lü et al. 2007), where ‘Kmax’ is the maximum desorbable amount of the element, ‘Kt’ is the cumulative amount of the element released over time ‘t’, and ‘t1/2’ is the half the time required to release the maximum amount.
}
\examples{
with(data = df_release, rk_second(Kt = Kt, Time = Time, ylab = "t/Kt"))
}
\references{
Grimme, H. 1980. The effect of field strength on quantity of K desorbed from soils by electro-ultrafiltration. Zeitschrift für Pflanzenernährung und Bodenkunde 143, 98–106. https://doi.org/10.1002/jpln.19801430113

Lü, X., Xu, J., Ma, W., Lu, Y., 2007. Comparison of seven kinetic equations for k release and application of kinetic parameters. Pedosphere 17, 124–129. https://doi.org/10.1016/S1002-0160(07)60017-4
}
