% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Release_kinetics.R
\name{rk_zero}
\alias{rk_zero}
\title{Fitting of release kinetics data to the zero-order equation}
\usage{
rk_zero(Kt = Kt, Time = Time,...)
}
\arguments{
\item{Kt}{Cumulative amount of element released (mass or mole of element per unit mass of soil, e.g., mg/kg or mmol/kg) after time ‘t’}

\item{Time}{The time intervals used in the kinetics study (user can choose the unit as per convenience)}

\item{...}{Any other argument that can be passed to base plot}
}
\value{
\itemize{
\item R2: Coefficient of determination (more the value, better the fit)
\item SEE: Standard error of estimate (less the value, better the fit)
\item a: Equation constant
\item b: Equation constant (overall release rate coefficient)
}
}
\description{
The function fits the cumulative amount of an element released from soil with a given extractant over a certain time period to the zero-order equation. It generates the coefficient of determination (R-squared) and standard error of estimate to show the goodness of fit. The parameters of zero-order equation, i.e., ‘a’ and ‘b’ can be obtained from this function.
}
\details{
The zero-order equation is expressed as (Km – Kt ) = a – b × t (Martin and Sparks 1983), where ‘Km’ is the total amount of the element released over the entire study period, ‘Kt’ is the cumulative amount of the element released over time ‘t’, and ‘a’ and ‘b’ are constants. The constant ‘b’ can be considered as the overall release rate coefficient.
}
\examples{
with(data = df_release, rk_zero(Kt = Kt, Time = Time, ylab = "Kmax - Kt"))
}
\references{
Martin, H.W., Sparks, D.L., 1983. Kinetics of nonexchangeable potassium release from two coastal plain soils. Soil Science Society of America Journal 47, 883–887. https://doi.org/10.2136/sssaj1983.03615995004700050008x
}
